requirements.txt
# Python version
python==3.11

absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.4
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.4.0
optree==0.13.0
pillow==11.0.0
protobuf==4.25.5
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.2.0
python-multipart==0.0.17
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
urllib3==2.2.3
uvicorn==0.32.0
werkzeug==3.1.3
wrapt==1.16.0
